[nfc] Resolve incompatible type comparison warnings
The recent changes from PR #723 cause a handful of comparisons in the
code to trigger warnings. None of them are actual problems, and this
patch disambiguates the comparisons to remove the warning.

+++ lib/Targets/AIEVecToCpp/TranslateAIEVecToCpp.cpp
@@ -348,7 +348,8 @@ static LogicalResult createLinearizedAccess(CppEmitter &emitter, Value source,
   ArrayRef<int64_t> stride = memRefType.getShape();
 
   // The stride and indices size must match
-  if (stride.size() != indices.size() || stride.size() != memRefType.getRank())
+  if (stride.size() != indices.size() ||
+      static_cast<int64_t>(stride.size()) != memRefType.getRank())
     return failure();
 
   // A stride contains two parts:
